changed size_t to int in matrix.cpp

+++ src/libs/matrix/matrix.cpp
@@ -71,7 +71,7 @@ void Matrix::AddMatrix(const Matrix& MatrixB) {
 	int row;
 #pragma omp parallel for
 	for (row = 0; row < this->rows; row++) {
-		for (size_t col = 0; col < this->cols; col++) {
+		for (int col = 0; col < this->cols; col++) {
 			this->data[row][col] += MatrixB.data[row][col];
 		}
 	}
@@ -92,8 +92,8 @@ Matrix Matrix::MultiplyMatricesParallel(const Matrix& MatrixA, const Matrix& Mat
 	int rowA;
 #pragma omp parallel for
 	for (rowA = 0; rowA < MatrixA.rows; rowA++) {
-		for (size_t colA_rowB = 0; colA_rowB < MatrixA.cols; colA_rowB++) {
-			for (size_t colB = 0; colB < MatrixB.cols; colB++) {
+		for (int colA_rowB = 0; colA_rowB < MatrixA.cols; colA_rowB++) {
+			for (int colB = 0; colB < MatrixB.cols; colB++) {
 				data[rowA][colB] += MatrixA.data[rowA][colA_rowB] * MatrixB.data[colA_rowB][colB];
 			}
 		}
@@ -113,7 +113,7 @@ Matrix Matrix::VectorToMatrix(const vector<double>& vect, bool transposeVector) 
 		return Matrix(vector<vector<double>>(1, vect));
 	}
 	vector<vector<double>> data(vect.size(), vector<double>(1));
-	for (size_t i = 0; i < vect.size(); i++) {
+	for (int i = 0; i < vect.size(); i++) {
 		data[i][0] = vect[i];
 	}
 	return Matrix(data);
@@ -137,7 +137,7 @@ vector<double> Matrix::MultiplyMatrixByVector(const Matrix& MatrixA, const vecto
 	int rowA;
 #pragma omp parallel for
 	for (rowA = 0; rowA < MatrixA.rows; rowA++) {
-		for (size_t colA_rowB = 0; colA_rowB < VectorB.size(); colA_rowB++) {
+		for (int colA_rowB = 0; colA_rowB < VectorB.size(); colA_rowB++) {
 			data[rowA] += MatrixA.data[rowA][colA_rowB] * VectorB[colA_rowB];
 		}
 		if (addBias) {
@@ -160,7 +160,7 @@ Matrix Matrix::MultiplyVectors(const vector<double>& VectorA, const vector<doubl
 	int row;
 #pragma omp parallel for
 	for (row = 0; row < VectorA.size(); row++) {
-		for (size_t col = 0; col < VectorB.size(); col++) {
+		for (int col = 0; col < VectorB.size(); col++) {
 			newMatrix.data[row][col] = VectorA[row] * VectorB[col];
 		}
 		if (addBias) {
@@ -176,9 +176,9 @@ Matrix Matrix::RandomMatrixSetSize(size_t rows, size_t cols, double variance) {
 	used for weights initialization
 	*/
 	vector<vector<double>> data;
-	for (size_t i = 0; i < rows; i++) {
+	for (int i = 0; i < rows; i++) {
 		vector<double> tmp;
-		for (size_t j = 0; j < cols; j++) {
+		for (int j = 0; j < cols; j++) {
 			double randNum = normalRandom();
 			tmp.push_back(randNum * variance);
 		}
